Closing a few issues.

+++ app/components/course-templates/edit.jsx
@@ -55,15 +55,18 @@ export default React.createClass({
 		this.history.pushState(null,`lesson/${classroomId}/${sectionId}/new`);
 	},
 	deleteSection(sectionId,e){
-		coursesData.removeSectionFromCourse(this.props.params.templateId, sectionId).then(res=>{
-			let newSections = (this.state.sections).filter((obj)=>{
-				return obj._id !== sectionId;
-			});
-			this.setState({
-				sections: newSections
-			});
+		let deleteConfirm = confirm('Are you sure you want to delete this section');
+		if(deleteConfirm) {
+			coursesData.removeSectionFromCourse(this.props.params.templateId, sectionId).then(res=>{
+				let newSections = (this.state.sections).filter((obj)=>{
+					return obj._id !== sectionId;
+				});
+				this.setState({
+					sections: newSections
+				});
 
-		});
+			});
+		}
 	},
 	renderSections(key, index){
 		return <li key={index} className="lessonGroup" id={this.state.sections[index]._id}>
@@ -82,9 +85,12 @@ export default React.createClass({
 				</li>
 	},
 	deleteTemplate(){
-		coursesData.deleteCourse(this.props.params.templateId).then(res=>{
-			this.history.pushState(null,`course-templates`);
-		});
+		let deleteConfirm = confirm('Are you sure you want to delete this template?');
+		if(deleteConfirm) {
+			coursesData.deleteCourse(this.props.params.templateId).then(res=>{
+				this.history.pushState(null,`course-templates`);
+			});
+		}
 	},
 	render() {
 		let lessons = this.state.course.lessons;
